chore(car): move car/ui to car/presentation/ui
