+++ src/app/(dashboard)/dashboard/cars/page.tsx
@@ -1,4 +1,4 @@
-import { CarsGallery } from '@/car/ui/cars-gallery/cars-gallery';
+import { CarsGallery } from '@/car/presentation/ui/cars-gallery/cars-gallery';
 import { DashboardMain } from '@/dashboard/ui/main/main';
 
 export default function CarsPage() {
