+++ src/module/car/presentation/ui/cards/details/data-field/data-field.tsx
