+++ src/module/car/presentation/ui/cards/details/expiration-row/expiration-row.tsx
