+++ src/module/car/presentation/ui/forms/add/add.tsx
@@ -1,4 +1,4 @@
-import { CarForm } from '@/car/ui/form/form';
+import { CarForm } from '@/car/presentation/ui/form/form';
 
 import { useAddForm } from './use-add';
