+++ src/module/car/presentation/ui/image/image.tsx
