+++ src/module/car/presentation/ui/modals/add/add.tsx
@@ -1,24 +1,22 @@
 import type { RefObject } from 'react';
 
-import { EditForm } from '@/car/ui/forms/edit/edit';
-import type { Car } from '@/types';
+import { AddForm } from '@/car/presentation/ui/forms/add/add';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
 
-interface EditModalProps {
-  car?: Car;
+type AddModalProps = {
   ref?: RefObject<DialogModalRef | null>;
   onSubmit?: () => void;
-}
+};
 
-export function EditModal({ ref, car, onSubmit }: EditModalProps) {
+export function AddModal({ ref, onSubmit }: AddModalProps) {
   return (
     <DialogModal ref={ref}>
       <DialogModal.Root>
-        <DialogModal.Heading>Edit car</DialogModal.Heading>
+        <DialogModal.Heading>Add car</DialogModal.Heading>
         <TextSeparator className="my-4" />
-        <EditForm car={car} onSubmit={onSubmit} />
+        <AddForm onSubmit={onSubmit} />
       </DialogModal.Root>
     </DialogModal>
   );
