+++ src/module/car/presentation/ui/modals/edit/edit.tsx
@@ -1,22 +1,24 @@
 import type { RefObject } from 'react';
 
-import { AddForm } from '@/car/ui/forms/add/add';
+import { EditForm } from '@/car/presentation/ui/forms/edit/edit';
+import type { Car } from '@/types';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
 
-type AddModalProps = {
+interface EditModalProps {
+  car?: Car;
   ref?: RefObject<DialogModalRef | null>;
   onSubmit?: () => void;
-};
+}
 
-export function AddModal({ ref, onSubmit }: AddModalProps) {
+export function EditModal({ ref, car, onSubmit }: EditModalProps) {
   return (
     <DialogModal ref={ref}>
       <DialogModal.Root>
-        <DialogModal.Heading>Add car</DialogModal.Heading>
+        <DialogModal.Heading>Edit car</DialogModal.Heading>
         <TextSeparator className="my-4" />
-        <AddForm onSubmit={onSubmit} />
+        <EditForm car={car} onSubmit={onSubmit} />
       </DialogModal.Root>
     </DialogModal>
   );
