+++ src/module/car/presentation/ui/sections/delete/use-delete.ts
