+++ src/module/car/presentation/ui/sections/details/controls/controls.tsx
@@ -1,7 +1,7 @@
 import { useRef } from 'react';
 
-import { EditButton } from '@/car/ui/buttons/edit/edit';
-import { EditModal } from '@/car/ui/modals/edit/edit';
+import { EditButton } from '@/car/presentation/ui/buttons/edit/edit';
+import { EditModal } from '@/car/presentation/ui/modals/edit/edit';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 import type { Car } from '@/types';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
