+++ src/module/car/presentation/ui/tables/date-expiration/placeholder/placeholder.tsx
